# OpenAPI.yaml
openapi: 3.1.0
info: 
  title: Course Registration System - OpenAPI 3.1
  description: |-
    This is the Course Registration System developed by Group Delta for the CMSC495 Computer Science Capstone Project.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://127.0.0.1:3000
...
paths:
...
  /admin:
    get:
      tags:
        - admin
      summary: List users
      description: Return a paginated list of users. Requires role ADMIN. Role is verified server-side for the session user.
      operationId: getUsers
      parameters:
        - name: page
          in: query
          description: Page number, default = 1
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          description: Items per page, default is 10
          required: true
          schema:
            type: integer
        - name: search
          in: query
          description: Search term
          required: false
          schema:
            type: string
        - name: role
          in: query
          description: "Role filter: ADMIN, STUDENT, or PROFESSOR"
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated user list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeta'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserMeta'
        '400':
          description: Invalid role filter
        '401':
          description: Invalid or missing session
        '403':
          description: "Forbidden: ADMIN role required"
      security:
        - SessionAuth: []
    post:
      tags:
        - admin
      summary: Create a user
      description: Create a new user account. Requires role ADMIN. Role is verified server-side for the session user.
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserCreateRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/AdminUserCreateRequest'
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid query parameter
        '401':
          description: Invalid or missing session
        '403':
          description: "Forbidden: ADMIN role required"
        '409':
          description: User with this email already exists
      security:
        - SessionAuth: []
...
  /course:
    get:
      tags:
        - course
      summary: List courses
      description: Return a paginated list of courses.
      operationId: getCourses
      parameters:
        - name: page
          in: query
          description: Page number, default = 1
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          description: Items per page, default is 10
          required: true
          schema:
            type: integer
        - name: search
          in: query
          description: Search term
          required: false
          schema:
            type: string
        - name: subject
          in: query
          description: Subject code filter
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated course list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseMeta'
            application/xml:
              schema:
                $ref: '#/components/schemas/CourseMeta'
        '400':
          description: Invalid query parameters
    post:
      tags:
        - course
      summary: Create a course
      description: Create a new course. Requires role ADMIN. Role is verified server-side for the session user.
      operationId: addCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseWriteRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/CourseWriteRequest'
      responses:
        '201':
          description: Course created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
            application/xml:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Invalid request body
        '401':
          description: Invalid or missing session
        '403':
          description: "Forbidden: ADMIN role required"
        '409':
          description: Course with this course code already exists
        '500':
          description: Failed to retrieve newly created course
      security:
        - SessionAuth: []
...
  /prerequisite:
    post:
      tags:
        - prerequisite
      summary: Create a prerequisite
      description: Add a prerequisite relationship to a course. Requires role ADMIN. Role is verified server-side for the session user.
      operationId: addPrerequisite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrerequisiteCreateRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/PrerequisiteCreateRequest'
      responses:
        '201':
          description: Prerequisite created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prerequisite'
            application/xml:
              schema:
                $ref: '#/components/schemas/Prerequisite'
        '400':
          description: Invalid request body or a course cannot be its own prerequisite
        '401':
          description: Invalid or missing session
        '403':
          description: "Forbidden: ADMIN role required"
        '404':
          description: One or more courses not found
        '409':
          description: Prerequisite relationship already exists or would create a cycle
        '500':
          description: Failed to add prerequisite relationship
      security:
        - SessionAuth: []
...
  /semester:
    get:
      tags:
        - semester
      summary: List semesters
      description: Return all semesters.
      operationId: getSemesters
      parameters:
      responses:
        '200':
          description: Semester list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemesterArray'
            application/xml:
              schema:
                $ref: '#/components/schemas/SemesterArray'
    post:
      tags:
        - semester
      summary: Create a semester
      description: Create a new semester. Requires role ADMIN. Role is verified server-side for the session user.
      operationId: addSemester
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemesterWriteRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/SemesterWriteRequest'
      responses:
        '201':
          description: Semester created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Semester'
            application/xml:
              schema:
                $ref: '#/components/schemas/Semester'
        '400':
          description: Invalid request body
        '401':
          description: Invalid or missing session
        '403':
          description: "Forbidden: ADMIN role required"
        '409':
          description: Semester already exists
      security:
        - SessionAuth: []
...
  /section:
    get:
      tags:
        - section
      summary: List sections
      description: Return a paginated list of sections.
      operationId: getSections
      parameters:
        - name: page
          in: query
          description: Page number, default = 1
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          description: Items per page, default is 10
          required: true
          schema:
            type: integer
        - name: search
          in: query
          description: Search term
          required: true
          schema:
            type: string
        - name: crsId
          in: query
          description: Course ID, default = null
          required: false
          schema:
            type: integer
        - name: semId
          in: query
          description: Semester ID, default = null
          required: false
          schema:
            type: integer
        - name: profId
          in: query
          description: Professor ID, default = null
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Paginated section list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionMeta'
            application/xml:
              schema:
                $ref: '#/components/schemas/SectionMeta'
        '400':
          description: Invalid query parameters
...
  /enrollment:
    post:
      tags:
        - enrollment
      summary: Create an enrollment
      description: Create a new enrollment record. Requires role STUDENT or ADMIN. Role is verified server-side for the session user. Students may only manage their own enrollments.
      operationId: addEnrollment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentCreateRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/EnrollmentCreateRequest'
      responses:
        '201':
          description: Enrollment created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '400':
          description: Invalid request body or access code
        '401':
          description: Invalid or missing session
        '403':
          description: "Forbidden: STUDENT or ADMIN role required"
        '404':
          description: Student or section not found
        '409':
          description: Prerequisites not met, student already enrolled, or section is full
      security:
        - SessionAuth: []
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: sid
      description: HttpOnly session cookie containing a session identifier. The server resolves the user from the session and verifies role server-side for restricted operations.
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        email:
          type: string
          format: email
        first_login:
          type: boolean
        role:
          type: string
        role_id:
          type: integer
        role_details:
          type: string
...
    AdminUserCreateRequest:
      type: object
      required:
        - name
        - email
        - detail
        - type
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        detail:
          type: string
        type:
          type: string
          enum:
            - ADMIN
            - STUDENT
            - PROFESSOR
...
    UserArray:
      type: array
      items:
        type: object
        properties:
          User:
            $ref: '#/components/schemas/User'
    UserMeta:
      type: object
      properties:
        User:
          type: array
          $ref: '#/components/schemas/UserArray'
        Meta:
          type: object
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Course:
      type: object
      properties:
        course_id:
          type: integer
        course_code:
          type: string
        title:
          type: string
        description:
          type: string
        credits:
          type: integer
        subject:
          type: string
    CourseWriteRequest:
      type: object
      required:
        - code
        - title
        - desc
        - cred
      properties:
        code:
          type: string
        title:
          type: string
        desc:
          type: string
        cred:
          type: integer
    CourseArray:
      type: array
      items:
        type: object
        properties:
          Course:
            $ref: '#/components/schemas/Course'
    CourseMeta:
      type: object
      properties:
        Course:
          type: array
          $ref: '#/components/schemas/CourseArray'
        Meta:
          type: object
          $ref: '#/components/schemas/Meta'
    Prerequisite:
      type: object
      properties:
        courseId:
          type: integer
        courseCode:
          type: string
        title:
          type: string
    PrerequisiteCreateRequest:
      type: object
      required:
        - cId
        - pId
      properties:
        cId:
          type: integer
        pId:
          type: integer
...
    Semester:
      type: object
      properties:
        semester_id:
          type: integer
        term:
          type: string
        year:
          type: integer
    SemesterWriteRequest:
      type: object
      required:
        - term
        - year
      properties:
        term:
          type: string
        year:
          type: integer
    SemesterArray:
      type: array
      items:
        type: object
        properties:
          Semester:
            $ref: '#/components/schemas/Semester'
    Section:
      type: object
      properties:
        section_id:
          type: integer
        course_id:
          type: integer
        semester_id:
          type: integer
        professor_id:
          type: integer
        capacity:
          type: integer
        days:
          type: string
        start_time:
          type: string
          description: Time of day in HH:MM:SS format.
          pattern: '^\d{2}:\d{2}:\d{2}$'
        end_time:
          type: string
          description: Time of day in HH:MM:SS format.
          pattern: '^\d{2}:\d{2}:\d{2}$'
...
    SectionArray:
      type: array
      items:
        type: object
        properties:
          Section:
            $ref: '#/components/schemas/Section'
    SectionMeta:
      type: object
      properties:
        Section:
          type: array
          $ref: '#/components/schemas/SectionArray'
        Meta:
          type: object
          $ref: '#/components/schemas/Meta'
    Enrollment:
      type: object
      properties:
        enrollment_id:
          type: integer
        student_id:
          type: integer
        section_id:
          type: integer
        status:
          type: string
          enum:
            - enrolled
            - waitlisted
            - dropped
            - completed
    EnrollmentCreateRequest:
      type: object
      required:
        - secId
        - stuId
        - code
      properties:
        secId:
          type: integer
        stuId:
          type: integer
        code:
          type: string
          description: Access code required to enroll in the section.
...